Move NgoList component to it's own folder and remove old ngo list route
#97
