+++ app/javascript/components/App.js
@@ -2,8 +2,6 @@ import React from "react"
 import {BrowserRouter, Switch, Route} from "react-router-dom";
 import {Provider} from 'react-redux';
 import configureStore from '../configureStore';
-import NgosList from "../containers/NgosList";
-import NgoPage from "../containers/NgoPage";
 import ErrorBoundary from "./ErrorBoundary";
 import AdoptionList from "../containers/AdoptionList/AdoptionList";
 import Navigation from "./Navigation/Navigation";
@@ -11,6 +9,8 @@ import SideNavigation from "./Navigation/SideNavigation/SideNavigation";
 import HomePage from "../containers/HomePage/HomePage";
 import DialogMessage from "./DialogMessage/DialogMessage";
 import UserSettings from "../containers/UserSettings/UserSettings";
+import NgoList from "../containers/NgoList/NgoList";
+import NgoPage from "../containers/NgoList/NgoPage";
 
 import '../fonts/roboto/index.css';
 
@@ -43,9 +43,8 @@ class App extends React.Component {
                         />
                         <Switch>
                             <Route exact path="/" component={HomePage} />
-                            <Route exact path="/ongs" render={() => window.location.href = '/ongs'} />
-                            <Route path="/new/ongs" render={() => <NgosList />} />
-                            <Route exact path="/new/ong/:id" component={NgoPage} />
+                            <Route path="/ongs" component={NgoList} />
+                            <Route exact path="/ong/:id" component={NgoPage} />
                             <Route path="/adocao" component={AdoptionList} />
                             <Route path="/settings" component={UserSettings} />
                         </Switch>
