+++ app/javascript/containers/NgoList/NgoList.sass
@@ -1,0 +1,1 @@
+.NgoList
